whatsapp-api/pyproject.toml
[project]
name = "whatsapp-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Madhav Sharma",email = "madhav-sharma@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "greenlet (>=3.2.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "whatsapp_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
